weather_writer/requirements.txt
fastapi
kafka-python
pymongo
python-dotenv
uvicorn
